Added working support for blob storage in cassandra. Previously blobs were being correctly serialized as a json object with the format: {"type": "Buffer", "data": [<buffer data>]} But on de-serialization the json object was being passed directly to the cassandra driver instead of a Buffer object being constructed from the json object.
I had get rid of the batch processing to achieve this as even small
(1-2MB) blobs can cause a single query to exceed the batch size limit.
Since batches should only be used for multi-table queries that require
synchronization (a situation which is impossible in this context) or for
single partition multi-query batches (a situation which is very unlikely
in this context), this isn't a problem. cf.
https://docs.datastax.com/en/cql/3.3/cql/cql_using/useBatchGoodExample.html

Modified import.js and export.js to support PORT environment variable as
it was convenient during debugging.